requirements.txt
asgiref==3.4.1
attrs==21.2.0
Automat==20.2.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
constantly==15.1.0
cryptography==36.0.0
cssselect==1.1.0
Django==3.2.9
django-elasticsearch-dsl==7.2.1
elasticsearch==7.16.1
elasticsearch-dsl==7.4.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.4.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.6.4
parsel==1.6.0
Pillow==8.4.0
priority==1.3.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyDispatcher==2.0.5
pyOpenSSL==21.0.0
python-dateutil==2.8.2
pytz==2021.3
queuelib==1.6.2
Scrapy==2.5.1
scrapy-djangoitem==1.1.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
Twisted==21.7.0
typing-extensions==4.0.0
urllib3==1.26.7
w3lib==1.22.0
zope.interface==5.4.0
